# internal/domain/models/NoteModels.go
package models

import (
	"github.com/google/uuid"
	"time"
)

type Note struct {
	ID        uuid.UUID `json:"id"`
	UserId    uuid.UUID `json:"user_id"`
	Title     string    `json:"titel"`
	Content   string    `json:"content"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}